feat: add save_recording method to BaseEnv for saving recorded states
